requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
accelerate=0.24.1=pypi_0
astunparse=1.6.3=pypi_0
beautifulsoup4=4.12.2=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
deepface=0.0.79=pypi_0
diffusers=0.23.0=pypi_0
filelock=3.13.1=pypi_0
fire=0.5.0=pypi_0
flask=3.0.0=pypi_0
flatbuffers=23.5.26=pypi_0
fsspec=2023.10.0=pypi_0
gast=0.5.4=pypi_0
gdown=4.7.1=pypi_0
google-auth=2.23.4=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.59.2=pypi_0
gunicorn=21.2.0=pypi_0
h5py=3.10.0=pypi_0
huggingface-hub=0.17.3=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
keras=2.14.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.5.1=pypi_0
markupsafe=2.1.3=pypi_0
ml-dtypes=0.2.0=pypi_0
mpmath=1.3.0=pypi_0
mtcnn=0.1.1=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
numpy=1.26.1=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.52=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
openssl=3.0.12=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
pandas=2.1.2=pypi_0
pillow=10.1.0=pypi_0
pip=23.3=py311h06a4308_0
protobuf=4.25.0=pypi_0
psutil=5.9.6=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pysocks=1.7.1=pypi_0
python=3.11.5=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
retina-face=0.0.13=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.0=pypi_0
setuptools=68.0.0=py311h06a4308_0
six=1.16.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tensorboard=2.14.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.14.0=pypi_0
tensorflow-estimator=2.14.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=2.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.14.1=pypi_0
torch=2.1.0=pypi_0
tqdm=4.66.1=pypi_0
transformers=4.35.0=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.0.7=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311h06a4308_0
wrapt=1.14.1=pypi_0
xformers=0.0.22.post7=pypi_0
xz=5.4.2=h5eee18b_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
